refactor(ui): 弹窗外壳提为共享组件 XDialogShell
_FilterDialogShell 已被 4 个弹窗复用,但作为私有类无法跨文件使用。
提取到 shared/components 并按项目 X 前缀惯例命名,供即将新增的
主题色弹窗复用。纯重构,无行为变化。
